+++ lib/ex_pipedrive.ex
@@ -17,6 +17,8 @@ defmodule ExPipedrive do
   defdelegate get_organization(client, org_id), to: ExPipedrive.Organizations
   defdelegate get_person(client, person_id), to: ExPipedrive.Persons
   defdelegate list_activities(client, opts \\ []), to: ExPipedrive.Activities
+  defdelegate list_activities_page(client, opts \\ []), to: ExPipedrive.Activities
+  defdelegate stream_activities(client, opts \\ []), to: ExPipedrive.Activities
   defdelegate list_activity_types(client), to: ExPipedrive.ActivityTypes
   defdelegate list_deals(client, opts), to: ExPipedrive.Deals
   defdelegate list_deals_page(client, opts \\ []), to: ExPipedrive.Deals
@@ -25,6 +27,8 @@ defmodule ExPipedrive do
   defdelegate list_leads(client, opts \\ []), to: ExPipedrive.Leads
   defdelegate list_notes(client, opts \\ []), to: ExPipedrive.Notes
   defdelegate list_organizations(client, opts), to: ExPipedrive.Organizations
+  defdelegate list_organizations_page(client, opts \\ []), to: ExPipedrive.Organizations
+  defdelegate stream_organizations(client, opts \\ []), to: ExPipedrive.Organizations
   defdelegate list_organization_fields(client, opts), to: ExPipedrive.OrganizationFields
   defdelegate list_own_activities(client, opts \\ []), to: ExPipedrive.Activities
   defdelegate list_person_fields(client, opts), to: ExPipedrive.PersonFields